fmr_ptr() type added

+++ carbon/atmegau2/system.c
@@ -10,7 +10,7 @@ fmr_return fmr_push(struct _fmr_push_pull_packet *packet) {
 		return -1;
 	}
 	megausb_bulk_receive(swap, packet->length);
-	*(uintptr_t *)(packet->call.parameters) = (uintptr_t)swap;
+	*(uint32_t *)(packet->call.parameters) = (uint32_t)swap;
 	retval = fmr_execute(packet->call.index, packet->call.function, packet->call.argc, packet->call.types, (void *)(packet->call.parameters));
 	free(swap);
 	return retval;
@@ -23,7 +23,7 @@ fmr_return fmr_pull(struct _fmr_push_pull_packet *packet) {
 		lf_error_raise(E_MALLOC, NULL);
 		return -1;
 	}
-	*(uintptr_t *)(packet->call.parameters) = (uintptr_t)swap;
+	*(uint32_t *)(packet->call.parameters) = (uint32_t)swap;
 	retval = fmr_execute(packet->call.index, packet->call.function, packet->call.argc, packet->call.types, (void *)(packet->call.parameters));
 	megausb_bulk_transmit(swap, packet->length);
 	free(swap);
